@erikgrinaker 2022-08-21
I ran a few CPU profiles here to find out why a single range key adds ~25% overhead. It seemed to all be down in Pebble, specifically `interleavingIter`.

For 23.1, we could reconsider whether we need interleaved exports, and instead export point/range keys separately (which would eliminate any penalty). However, we'll likely have to improve the `interleavingIter` performance for e.g. the scan case anyway, so it may be moot.